added basic block render test && updated snapshots
